Rename gotrue, step 2.
