+++ packages/auth/src/authClients/goTrue.ts
