+++ packages/auth/src/authClients/index.ts
@@ -1,13 +1,13 @@
 import type { NetlifyIdentity } from './netlify'
 import type { Auth0, Auth0User } from './auth0'
-import type { GoTrue, GoTrueUser } from './gotrue'
+import type { GoTrue, GoTrueUser } from './goTrue'
 import type { MagicLink, MagicUser } from './magicLink'
 import type { Firebase } from './firebase'
 import type { Custom } from './custom'
 //
 import { netlify } from './netlify'
 import { auth0 } from './auth0'
-import { goTrue } from './gotrue'
+import { goTrue } from './goTrue'
 import { magicLink } from './magicLink'
 import { firebase } from './firebase'
 import { custom } from './custom'
